Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 int main()
 {
     const double pi = 3.14;
-    int radius;
+    double radius;
     double area;
 
     scanf("%d",&radius);
